requirements.txt
requests==2.28.1
python-rofi==1.0.1
